pyproject.toml
[project]
name = "personal-agent"
version = "0.1.0"
description = "Multi-agent personal assistant"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0.0",
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-openai>=1.1.0",
    "deepagents>=0.4.0",
    "langsmith>=0.7.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "Pillow>=10.0",
    "pillow-heif>=1.0",
    "pypdf>=4.0",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0",
    "watchdog>=4.0",
]

[project.scripts]
receipt-watcher = "agents.receipt_analyzer.watcher:run_watcher"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["shared*", "agents*", "orchestrator*", "ui*"]

[tool.ruff]
line-length = 100
